[chore] Timestamp int/double -> int64_t

+++ libdevcore/addrStore.h
@@ -31,12 +31,12 @@ namespace c2pool::dev
     struct AddrValue
     {
         int service;
-        int first_seen;
-        int last_seen;
+        int64_t first_seen;
+        int64_t last_seen;
 
         AddrValue() {}
 
-        AddrValue(int _service, int _first_seen, int _last_seen){
+        AddrValue(int _service, int64_t _first_seen, int64_t _last_seen){
             service = _service;
             first_seen = _first_seen;
             last_seen = _last_seen;
